+++ cypress/tests/core/objectBrowser.js
@@ -37,7 +37,7 @@ describe('Object Browser Tests', () => {
     cy.get('.toolbar-inner button.ui.basic.icon.button').click();
     cy.findByLabelText('Search SVG').click();
     cy.get('.ui.input.search').type('/my-page-1');
-    cy.findByLabelText('Select my-image').dblclick();
+    cy.findByLabelText('Select My Image').dblclick();
     cy.get('#toolbar-save').click();
     cy.url().should('eq', Cypress.config().baseUrl + '/my-page');
 
@@ -54,7 +54,7 @@ describe('Object Browser Tests', () => {
     cy.get('.toolbar-inner button.ui.basic.icon.button').click();
     cy.findByLabelText('Search SVG').click();
     cy.get('.ui.input.search').type('http://localhost:55001/my-page-1');
-    cy.findByLabelText('Select my-image').dblclick();
+    cy.findByLabelText('Select My Image').dblclick();
     cy.get('#toolbar-save').click();
     cy.url().should('eq', Cypress.config().baseUrl + '/my-page');